pyproject.toml
[tool.poetry]
name = "myListener"
version = "0.1.0"
description = ""
authors = ["TomW15 <58078501+TomW15@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
myDb = { git = "https://github.com/TomW15/myDb.git", branch = "master"}
numpy = "1.24.4"
pandas = "1.5.3"
pushbullet-py = "^0.12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[tool.poetry.scripts]
listen_to_trading_db = "myListener.listen:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
